requirements.txt
loguru==0.6.0
python-decouple==3.8 
fake_useragent==1.1.1
selenium==4.8.2
